services/ai-service/requirements.txt
agno>=2.1.2
openai>=1.88.0
chromadb>=1.0.13
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
httpx>=0.28.0
sqlalchemy>=2.0.0
